requirements.txt
torch>=2.3.0
torchaudio>=2.3.0
transformers>=4.42.0
PyYAML>=6.0.1
tqdm>=4.66.0
numpy>=1.26.0

